+++ uniffi_bindgen/src/bindings/swift/gen_swift/object.rs
@@ -2,25 +2,33 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::backend::CodeType;
+use crate::{backend::CodeType, interface::ObjectImpl};
 
 #[derive(Debug)]
 pub struct ObjectCodeType {
-    id: String,
+    name: String,
+    imp: ObjectImpl,
 }
 
 impl ObjectCodeType {
-    pub fn new(id: String) -> Self {
-        Self { id }
+    pub fn new(name: String, imp: ObjectImpl) -> Self {
+        Self { name, imp }
     }
 }
 
 impl CodeType for ObjectCodeType {
     fn type_label(&self) -> String {
-        super::SwiftCodeOracle.class_name(&self.id)
+        super::SwiftCodeOracle.class_name(&self.name)
     }
 
     fn canonical_name(&self) -> String {
-        format!("Type{}", self.id)
+        format!("Type{}", self.name)
+    }
+
+    fn initialization_fn(&self) -> Option<String> {
+        match &self.imp {
+            ObjectImpl::Struct => None,
+            ObjectImpl::Trait => Some(format!("uniffiCallbackInit{}", self.name)),
+        }
     }
 }